Big improvement with WimC's Bhaskara-oid approx.
This is a sine approximation in the style of Bhaskara I's formula, https://en.wikipedia.org/wiki/Bhaskara_I%27s_sine_approximation_formula , but slightly improved thanks to WimC's magic on the Mathematics StackExchange. It was tricky to get right, but it does seem quite high-quality; we'll need to see about speed, though.