We can avoid calls to floor() or fastFloor().
By treating a cast from float or double to int as just a typical truncation, we get the same general result from `d -= (int)d;` as `d %= 1.0`, that is, sometimes negative, sometimes positive. We can then use similar logic to the modulus-based degrees and radians code to make sure the angle is positive.