Check the "Nick" approx again vs. Bhaskara I.
Both are non-tabular and work with either float or double math. Bhaskara I (TrigTools.sinSmooth()) definitely is more accurate, both for its max abs error and all average error types. But, OldNumberTools.sin() is potentially needed to reproduce some results in SquidLib, as inaccurate as they may be...

This also updates Gradle and libGDX, to 8.2 and 1.12.0, respectively.