Should have run this test sooner...
InterpolationsTest is from SquidSquad (that's also where Interpolations is from), and I thought it was passing, but apparently not. It compares each Interpolation in libGDX to its counterpart in my code, and tries to ensure they produce the same or very similar output. It currently fails on the "bounce" Interpolator, and I'm not sure why. I already corrected elasticIn and elasticOut, which previously jumped from an opposite extreme to suddenly be at the correct point for the endpoints, with no time in the middle. That's a debatable bug that's still in libGDX.