While we're at it...
(It being breaking seeds), we can switch to the better randomize3() for the Hasher(CharSequence) constructor that we use a lot. I also don't think the hash() and hash64() methods that take a seed parameter need to partly-randomize it themselves; users can call randomize3(long) on a seed if they want to and get a more-random seed than before, and if they keep using that seed it doesn't need to be partly-randomized over and over again.