I wish I knew why mean rel. error does this.
For quite a lot of the tested sine functions, the mean relative error is 0.0154355764 . This doesn't seem to happen for anything that's a smooth function or that does piecewise interpolation, but does affect quite a lot of simple table lookups. The value 0.0154355764 doesn't seem to depend on the size of the table at all, which is strange.