Start on improving decimal support.
There are some lingering issues with the decimal code I added to Ryu, such as the input 0 not using padding. To help treat those issues, it's useful to have the full set of functions that use the decimal code. This commit adds joinDecimal() and appendJoinedDecimal(), which take length limit arguments.