Add some actual automated unit tests.
These check that the results for the TrigTools methods (most of them, anyway) are within a reasonable tolerance from the Math trig methods. The tan() variations don't check inputs close to HALF_PI above a multiple of PI, because those have very large outputs and a small relative difference is more than enough to break the absolute difference check.