Itty bitty quality improvements to Distributor.
Now, because the bitwise arcana isn't used for integer probit, numbers stay in the 0.0 to 1.0 range when we use a floating-point probit on them. We also add a much smaller double to probitL and probitD *after* squaring a value that could have been trouble if out-of-range before. Importantly, probitI and probitL consider an additional bit, doubling the possible values each can return.

This commit is very complicated and technical, but I updated the docs, so that should help a little.