Small optimizations in Distributor.
First, probitF() and probitI() don't actually need logRough, they can actually do fine with logRougher() -- the argument it takes is always over 460f anyway, so the precision on small values doesn't matter. Next, probitI() and probitL() don't need to add 0.5 at the start but subtract it 90% of the time in the last else block; they can lose less precision by staying close to 0 always, and the additions/subtractions are always the same in number or fewer. This also means probitI() and probitL() actually do consider the lowest bits as different inputs, and return different results for, say, 1 and 2 as inputs.