Avoid big calculations per-cbrt.
This helps speed up the testing process, which is needed because after 5 minutes this has completed only 2/5577 tests. Other optimizations have this not continue the innermost loop if it is going out of a local optimum; I'm not sure if this will help absolutely, but so far this seems to have one optimum per innermost loop. Without eliminating any, it would need to run 72501 tests...